main is not included

+++ includes/common.h
@@ -4,6 +4,6 @@
 
 extern size_t working_set_size;
 extern long long  working_set_per_thread;
-extern int zipfian_number;// common.h에 있을 전역 사용 변수 같은애들!
+extern double zipfian_number;// common.h에 있을 전역 사용 변수 같은애들!
 extern int number_of_threads;
 extern size_t memory_alloc_size;
